% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{adjust_dropout}
\alias{adjust_dropout}
\title{Adjust Sample Size for Dropouts}
\usage{
adjust_dropout(n, dropout_rate)
}
\arguments{
\item{n}{Required sample size}

\item{dropout_rate}{Anticipated dropout rate (proportion)}
}
\value{
Adjusted sample size
}
\description{
Inflate sample size to account for anticipated dropout rate
}
\examples{
adjust_dropout(n = 100, dropout_rate = 0.2)
}
