% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Bootstrap Confidence Interval}
\usage{
bootstrap_ci(data, statistic = mean, n_bootstrap = 1000, conf_level = 0.95)
}
\arguments{
\item{data}{Numeric vector}

\item{statistic}{Function to calculate statistic}

\item{n_bootstrap}{Number of bootstrap samples}

\item{conf_level}{Confidence level}
}
\value{
List with bootstrap results
}
\description{
Calculate bootstrap confidence interval
}
\examples{
data(omni_glucose)
bootstrap_ci(omni_glucose$hba1c_3m, mean)
}
