% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{bootstrap_regression}
\alias{bootstrap_regression}
\title{Bootstrap for Regression}
\usage{
bootstrap_regression(formula, data, n_bootstrap = 1000)
}
\arguments{
\item{formula}{Regression formula}

\item{data}{Data frame}

\item{n_bootstrap}{Number of bootstrap samples}
}
\value{
List with bootstrap results
}
\description{
Bootstrap confidence intervals for regression coefficients
}
\examples{
\dontrun{
bootstrap_regression(y ~ x, data)
}
}
