% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{calculate_km_for_plot}
\alias{calculate_km_for_plot}
\title{Internal KM calculation for plotting}
\usage{
calculate_km_for_plot(time, status)
}
\description{
Internal KM calculation for plotting
}
