% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{check_assumptions}
\alias{check_assumptions}
\title{Check Assumptions for Statistical Tests}
\usage{
check_assumptions(data, test = "ttest", group = NULL)
}
\arguments{
\item{data}{Data frame or numeric vector}

\item{test}{Type of test: "ttest", "anova", "regression"}

\item{group}{Grouping variable}
}
\value{
List with assumption check results
}
\description{
Check assumptions for common statistical tests
}
\examples{
data(omni_glucose)
check_assumptions(omni_glucose$hba1c_3m, "ttest",
                  group = omni_glucose$treatment)
}
