% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{cohen_d}
\alias{cohen_d}
\title{Cohen's d Effect Size}
\usage{
cohen_d(x, y)
}
\arguments{
\item{x}{Numeric vector for group 1}

\item{y}{Numeric vector for group 2}
}
\value{
List with Cohen's d and interpretation
}
\description{
Calculate Cohen's d for two independent groups
}
\examples{
data(omni_glucose)
cohen_d(omni_glucose$hba1c_baseline[1:100], omni_glucose$hba1c_baseline[101:200])
}
