% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{cohen_f}
\alias{cohen_f}
\title{Cohen's f for ANOVA}
\usage{
cohen_f(ss_between, ss_within, k = NULL)
}
\arguments{
\item{ss_between}{Between-group sum of squares}

\item{ss_within}{Within-group sum of squares}

\item{k}{Number of groups}
}
\value{
List with Cohen's f and interpretation
}
\description{
Calculate Cohen's f effect size for ANOVA
}
\examples{
cohen_f(ss_between = 100, ss_within = 400, k = 3)
}
