% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{convert_wide_to_long}
\alias{convert_wide_to_long}
\title{Convert Wide to Long Format}
\usage{
convert_wide_to_long(data, id_cols, names_to = "variable", values_to = "value")
}
\arguments{
\item{data}{Data frame in wide format}

\item{id_cols}{Columns to keep as identifiers}

\item{names_to}{Name of new column for variable names}

\item{values_to}{Name of new column for values}
}
\value{
Data frame in long format
}
\description{
Convert data from wide to long format
}
\examples{
data(omni_depression)
long_data <- convert_wide_to_long(
  omni_depression,
  id_cols = c("patient_id", "group"),
  names_to = "timepoint",
  values_to = "score"
)
}
