% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{create_dummy_vars}
\alias{create_dummy_vars}
\title{Create Dummy Variables}
\usage{
create_dummy_vars(data, vars, drop_first = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to convert}

\item{drop_first}{Drop first level (default TRUE)}
}
\value{
Data frame with dummy variables
}
\description{
Convert categorical variables to dummy/indicator variables
}
\examples{
data(omni_glucose)
create_dummy_vars(omni_glucose, vars = "treatment")
}
