% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{export_csv}
\alias{export_csv}
\title{Export CSV File}
\usage{
export_csv(data, file, row.names = FALSE, encoding = "UTF-8")
}
\arguments{
\item{data}{Data frame}

\item{file}{Output file path}

\item{row.names}{Include row names}

\item{encoding}{File encoding}
}
\description{
Export data frame to CSV file
}
\examples{
# export_csv(omni_glucose, "output.csv")
}
