% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{hazard_ratio}
\alias{hazard_ratio}
\title{Hazard Ratio}
\usage{
hazard_ratio(time, status, group)
}
\arguments{
\item{time}{Survival time}

\item{status}{Event status}

\item{group}{Binary grouping variable}
}
\value{
List with HR results
}
\description{
Calculate hazard ratio from survival data
}
\examples{
data(omni_cancer)
hazard_ratio(omni_cancer$time, omni_cancer$status, omni_cancer$treatment)
}
