% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV File}
\usage{
import_csv(file, header = TRUE, sep = ",", encoding = "UTF-8")
}
\arguments{
\item{file}{Path to CSV file}

\item{header}{Logical for header row}

\item{sep}{Separator character}

\item{encoding}{File encoding}
}
\value{
Data frame
}
\description{
Import data from CSV file with automatic encoding detection
}
\examples{
# import_csv("data.csv")
}
