% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge Datasets}
\usage{
merge_datasets(x, y, by = NULL, all = FALSE)
}
\arguments{
\item{x}{First data frame}

\item{y}{Second data frame}

\item{by}{Columns to merge by}

\item{all}{Logical for full outer join}
}
\value{
Merged data frame
}
\description{
Merge two data frames by common columns
}
\examples{
data(omni_glucose)
# Merge example (same data for demonstration)
merged <- merge_datasets(omni_glucose[1:10, ], omni_glucose[11:20, ], by = "patient_id")
}
