% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{multiple_imputation}
\alias{multiple_imputation}
\title{Multiple Imputation (Simple)}
\usage{
multiple_imputation(data, m = 5)
}
\arguments{
\item{data}{Data frame}

\item{m}{Number of imputations}
}
\value{
List of imputed data frames
}
\description{
Simple multiple imputation using predictive mean matching
}
\examples{
\dontrun{
imputed_data <- multiple_imputation(data, m = 5)
}
}
