% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{multiple_testing_correction}
\alias{multiple_testing_correction}
\title{Multiple Testing Correction}
\usage{
multiple_testing_correction(p_values, method = "fdr")
}
\arguments{
\item{p_values}{Vector of p-values}

\item{method}{Correction method: "bonferroni", "holm", "fdr", "bh"}
}
\value{
Data frame with original and adjusted p-values
}
\description{
Apply multiple testing correction methods
}
\examples{
p <- c(0.01, 0.03, 0.05, 0.1, 0.2)
multiple_testing_correction(p, method = "fdr")
}
