% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{nnt}
\alias{nnt}
\title{Number Needed to Treat}
\usage{
nnt(p_treatment, p_control)
}
\arguments{
\item{p_treatment}{Proportion with outcome in treatment group}

\item{p_control}{Proportion with outcome in control group}
}
\value{
List with NNT results
}
\description{
Calculate NNT with confidence interval
}
\examples{
nnt(p_treatment = 0.15, p_control = 0.25)
}
