% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_cancer}
\alias{omni_cancer}
\title{Cancer Treatment Outcomes Dataset}
\format{
A data frame with 300 rows and 9 variables:
\describe{
  \item{patient_id}{Patient identifier (CA0001-CA0300)}
  \item{age}{Age in years}
  \item{cancer_stage}{Cancer stage (I期/II期/III期/IV期)}
  \item{treatment}{Treatment modality (手术/化疗/放疗/免疫)}
  \item{ecog_score}{ECOG performance status (0-3)}
  \item{biomarker_level}{Biomarker level}
  \item{response}{Treatment response (无响应/有响应)}
  \item{time_to_event}{Survival time in months}
  \item{status}{Event status (1=event, 0=censored)}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_cancer
}
\description{
Cancer treatment outcomes for 300 patients including survival data.
Supports logistic regression and survival analysis (Kaplan-Meier, Cox PH).
}
\keyword{datasets}
