% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_cardio}
\alias{omni_cardio}
\title{Cardiovascular Risk Factors Dataset}
\format{
A data frame with 500 rows and 8 variables:
\describe{
  \item{patient_id}{Patient identifier (CV0001-CV0500)}
  \item{sbp}{Systolic blood pressure (mmHg)}
  \item{dbp}{Diastolic blood pressure (mmHg)}
  \item{ldl}{LDL cholesterol (mg/dL)}
  \item{age}{Age in years}
  \item{smoking_packyears}{Smoking pack-year history}
  \item{exercise_hours}{Weekly exercise hours}
  \item{cv_risk}{10-year cardiovascular risk score}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_cardio
}
\description{
Cardiovascular risk factors for 500 patients including blood pressure,
lipid profile, and lifestyle factors. Supports correlation analysis
and linear regression modeling.
}
\keyword{datasets}
