% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_depression}
\alias{omni_depression}
\title{Depression Repeated Measures Dataset}
\format{
A data frame with 1000 rows and 7 variables:
\describe{
  \item{patient_id}{Patient identifier (RM0001-RM0200)}
  \item{center}{Study center (中心1-中心4)}
  \item{time}{Time point label (第0周-第4周)}
  \item{time_num}{Numeric time (0-4)}
  \item{age}{Age in years}
  \item{depression_score}{Depression score (0-30)}
  \item{medication_adherence}{Medication adherence (0-100)}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_depression
}
\description{
Longitudinal depression scores for 200 patients measured at 5 time points
(1000 observations). Supports mixed-effects models and repeated measures ANOVA.
}
\keyword{datasets}
