% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_dose}
\alias{omni_dose}
\title{Dose-Response Dataset}
\format{
A data frame with 250 rows and 3 variables:
\describe{
  \item{subject_id}{Subject identifier (DR0001-DR0250)}
  \item{dose_mg}{Drug dose in mg (0, 5, 10, 20, 40, 80, 160, 320)}
  \item{response}{Physiological response (continuous variable)}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_dose
}
\description{
Dose-response relationship data for 250 subjects across 8 dose levels.
Supports non-linear regression, Emax model fitting, and dose-response analysis.
}
\keyword{datasets}
