% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_glucose}
\alias{omni_glucose}
\title{Glucose Control Study Dataset}
\format{
A data frame with 500 rows and 11 variables:
\describe{
  \item{patient_id}{Patient identifier (P0001-P0500)}
  \item{age}{Age in years}
  \item{gender}{Gender (男/女)}
  \item{bmi}{Body mass index (kg/m^2)}
  \item{diabetes_type}{Diabetes type (1型/2型/妊娠糖尿病)}
  \item{duration}{Disease duration in years}
  \item{hba1c_baseline}{Baseline HbA1c level (\%)}
  \item{treatment}{Treatment regimen (胰岛素/口服药/联合治疗)}
  \item{complications}{Complications (无/微血管/大血管/两者)}
  \item{fasting_glucose}{Fasting glucose (mmol/L)}
  \item{hba1c_3m}{HbA1c after 3 months (\%)}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_glucose
}
\description{
A dataset containing HbA1c and clinical parameters for 500 diabetic patients.
Covers t-test, ANOVA, and basic statistical analyses for diabetes research.
}
\keyword{datasets}
