% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{omni_ml}
\alias{omni_ml}
\title{Machine Learning Features Dataset}
\format{
A data frame with 350 rows and 24 variables:
\describe{
  \item{sample_id}{Sample identifier (ML0001-ML0350)}
  \item{age}{Age in years}
  \item{gender}{Gender (男/女)}
  \item{class}{Outcome class (稳定/进展)}
  \item{feature_1}{Feature variable 1}
  \item{feature_2}{Feature variable 2}
  \item{feature_3}{Feature variable 3}
  \item{feature_4}{Feature variable 4}
  \item{feature_5}{Feature variable 5}
  \item{feature_6}{Feature variable 6}
  \item{feature_7}{Feature variable 7}
  \item{feature_8}{Feature variable 8}
  \item{feature_9}{Feature variable 9}
  \item{feature_10}{Feature variable 10}
  \item{feature_11}{Feature variable 11}
  \item{feature_12}{Feature variable 12}
  \item{feature_13}{Feature variable 13}
  \item{feature_14}{Feature variable 14}
  \item{feature_15}{Feature variable 15}
  \item{feature_16}{Feature variable 16}
  \item{feature_17}{Feature variable 17}
  \item{feature_18}{Feature variable 18}
  \item{feature_19}{Feature variable 19}
  \item{feature_20}{Feature variable 20}
}
}
\source{
Modified from \url{https://gitee.com/ablikim-ali/data_for_learn/blob/master/clinical_stat_datasets.xlsx}
}
\usage{
omni_ml
}
\description{
Clinical features for 350 patients with 20 feature variables and binary outcome.
Supports random forest, SVM, and other ML classification algorithms.
}
\keyword{datasets}
