% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{plot_bland_altman}
\alias{plot_bland_altman}
\title{Bland-Altman Plot}
\usage{
plot_bland_altman(method1, method2, title = "Bland-Altman Plot")
}
\arguments{
\item{method1}{Measurements from method 1}

\item{method2}{Measurements from method 2}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Create Bland-Altman plot for method comparison
}
\examples{
\dontrun{
plot_bland_altman(method1, method2)
}
}
