% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Box Plot}
\usage{
plot_boxplot(x, y, title = NULL, xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{Grouping variable}

\item{y}{Numeric response variable}

\item{title}{Plot title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\value{
ggplot object
}
\description{
Create a box plot for comparing groups
}
\examples{
data(omni_glucose)
plot_boxplot(omni_glucose$treatment, omni_glucose$hba1c_3m,
             title = "HbA1c by Treatment Group")
}
