% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{plot_cumulative_incidence}
\alias{plot_cumulative_incidence}
\title{Cumulative Incidence Plot}
\usage{
plot_cumulative_incidence(
  time,
  status,
  group = NULL,
  title = "Cumulative Incidence"
)
}
\arguments{
\item{time}{Time to event}

\item{status}{Event status (0 = censored, 1 = event of interest, 2 = competing risk)}

\item{group}{Optional grouping variable}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot cumulative incidence function for competing risks
}
\examples{
\dontrun{
plot_cumulative_incidence(time, status, group)
}
}
