% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_complete.R
\name{plot_effect_size_curve}
\alias{plot_effect_size_curve}
\title{Plot Effect Size Curve}
\usage{
plot_effect_size_curve(
  effect_range = seq(0.1, 1, by = 0.1),
  n = 100,
  alpha = 0.05,
  test_type = "ttest"
)
}
\arguments{
\item{effect_range}{Range of effect sizes}

\item{n}{Sample size}

\item{alpha}{Significance level}

\item{test_type}{Type of test}
}
\value{
ggplot object
}
\description{
Visualize power across different effect sizes
}
\examples{
plot_effect_size_curve(effect_range = seq(0.1, 1, by = 0.1), n = 100)
}
