% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_complete.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Plot Power Curve}
\usage{
plot_power_curve(
  n_range = seq(10, 200, by = 10),
  effect_size = 0.5,
  alpha = 0.05,
  test_type = "ttest",
  ...
)
}
\arguments{
\item{n_range}{Range of sample sizes}

\item{effect_size}{Effect size}

\item{alpha}{Significance level (default 0.05)}

\item{test_type}{Type of test}

\item{...}{Additional parameters}
}
\value{
ggplot object
}
\description{
Visualize power across different sample sizes
}
\examples{
plot_power_curve(n_range = seq(10, 200, by = 10), effect_size = 0.5)
}
