% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC Curve (Alias)}
\usage{
plot_roc(predictor, response, title = "ROC Curve", ci = TRUE)
}
\arguments{
\item{predictor}{Predictor variable}

\item{response}{Response variable (binary)}

\item{title}{Plot title}

\item{ci}{Show confidence interval (default TRUE)}
}
\value{
List with AUC and plot
}
\description{
Alias for plot_roc_enhanced. See \code{\link{plot_roc_enhanced}} for details.
}
\examples{
data(omni_diagnostic)
disease_binary <- ifelse(omni_diagnostic$true_disease == "阳性", 1, 0)
plot_roc(omni_diagnostic$test1_result, disease_binary)
}
