% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{plot_swimmer}
\alias{plot_swimmer}
\title{Swimmer Plot}
\usage{
plot_swimmer(data, title = "Swimmer Plot")
}
\arguments{
\item{data}{Data frame with start, end, and event information}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Create swimmer plot for treatment duration
}
\examples{
\dontrun{
plot_swimmer(data)
}
}
