% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(
  effect_size,
  p_values,
  labels = NULL,
  fc_threshold = 1,
  p_threshold = 0.05,
  title = "Volcano Plot"
)
}
\arguments{
\item{effect_size}{Vector of effect sizes}

\item{p_values}{Vector of p-values}

\item{labels}{Optional labels}

\item{fc_threshold}{Fold change threshold}

\item{p_threshold}{P-value threshold}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Create volcano plot for differential analysis
}
\examples{
fc <- rnorm(100, 0, 2)
p <- runif(100, 0, 1)
plot_volcano(fc, p)
}
