% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_complete.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Waterfall Plot}
\usage{
plot_waterfall(response, group = NULL, title = "Waterfall Plot")
}
\arguments{
\item{response}{Change from baseline (percentage)}

\item{group}{Optional grouping variable}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Create waterfall plot for treatment response
}
\examples{
\dontrun{
plot_waterfall(response, group)
}
}
