% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis_complete.R
\name{power_cluster_randomized}
\alias{power_cluster_randomized}
\title{Power for Cluster Randomized Trial}
\usage{
power_cluster_randomized(
  n_clusters,
  cluster_size,
  delta,
  sd,
  icc,
  alpha = 0.05
)
}
\arguments{
\item{n_clusters}{Number of clusters per group}

\item{cluster_size}{Average cluster size}

\item{delta}{Effect size}

\item{sd}{Standard deviation}

\item{icc}{Intraclass correlation}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with power results
}
\description{
Calculate power for cluster randomized design
}
\examples{
power_cluster_randomized(n_clusters = 20, cluster_size = 30,
                         delta = 0.5, sd = 1, icc = 0.05)
}
