% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{relative_risk}
\alias{relative_risk}
\title{Relative Risk}
\usage{
relative_risk(a, b, c, d)
}
\arguments{
\item{a}{Exposed cases}

\item{b}{Exposed total}

\item{c}{Unexposed cases}

\item{d}{Unexposed total}
}
\value{
List with RR results
}
\description{
Calculate relative risk and confidence interval
}
\examples{
relative_risk(a = 30, b = 50, c = 15, d = 50)
}
