% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{ss_anova}
\alias{ss_anova}
\title{Sample Size for One-Way ANOVA}
\usage{
ss_anova(k, f, power = 0.8, alpha = 0.05)
}
\arguments{
\item{k}{Number of groups}

\item{f}{Cohen's f effect size}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Convenience wrapper for ANOVA sample size
}
\examples{
ss_anova(k = 3, f = 0.25)
}
