% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{ss_chisq}
\alias{ss_chisq}
\title{Sample Size for Chi-Square Test}
\usage{
ss_chisq(w, df = 1, power = 0.8, alpha = 0.05)
}
\arguments{
\item{w}{Cohen's w effect size}

\item{df}{Degrees of freedom}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Convenience wrapper for chi-square test sample size
}
\examples{
ss_chisq(w = 0.3, df = 1)
}
