% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_cluster_randomized}
\alias{ss_cluster_randomized}
\title{Cluster Randomized Trial Sample Size}
\usage{
ss_cluster_randomized(delta, sd, icc, cluster_size, power = 0.8, alpha = 0.05)
}
\arguments{
\item{delta}{Effect size}

\item{sd}{Standard deviation}

\item{icc}{Intraclass correlation coefficient}

\item{cluster_size}{Average cluster size}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for cluster randomized trial
}
\examples{
ss_cluster_randomized(delta = 0.5, sd = 1, icc = 0.05, cluster_size = 20)
}
