% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{ss_correlation}
\alias{ss_correlation}
\title{Sample Size for Correlation Test}
\usage{
ss_correlation(r, power = 0.8, alpha = 0.05, alternative = "two.sided")
}
\arguments{
\item{r}{Expected correlation coefficient}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}

\item{alternative}{"two.sided" or "one.sided"}
}
\value{
List with sample size results
}
\description{
Convenience wrapper for Pearson correlation sample size
}
\examples{
ss_correlation(r = 0.3)
}
