% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_cox_regression}
\alias{ss_cox_regression}
\title{Cox Regression Sample Size}
\usage{
ss_cox_regression(hr, r2 = 0, p_event = 0.5, power = 0.8, alpha = 0.05)
}
\arguments{
\item{hr}{Hazard ratio for the predictor of interest}

\item{r2}{R-squared from regression of predictor on other covariates}

\item{p_event}{Event probability (proportion experiencing event)}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for Cox proportional hazards regression
Based on Hsieh & Lavori (2000)
}
\examples{
ss_cox_regression(hr = 2.0, r2 = 0, p_event = 0.3)
}
