% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_group_sequential}
\alias{ss_group_sequential}
\title{Group Sequential Design Sample Size}
\usage{
ss_group_sequential(
  delta,
  sd,
  n_interim = 1,
  alpha_spending = "obf",
  power = 0.8,
  alpha = 0.05
)
}
\arguments{
\item{delta}{Effect size}

\item{sd}{Standard deviation}

\item{n_interim}{Number of planned interim analyses}

\item{alpha_spending}{Alpha spending function: "pocock" or "obf"}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Overall significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for group sequential trial
}
\examples{
ss_group_sequential(delta = 0.5, sd = 1, n_interim = 3)
}
