% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_mann_whitney}
\alias{ss_mann_whitney}
\title{Mann-Whitney U Test Sample Size}
\usage{
ss_mann_whitney(p, power = 0.8, alpha = 0.05)
}
\arguments{
\item{p}{Probability that observation from group 1 > observation from group 2}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for Mann-Whitney U test
}
\examples{
ss_mann_whitney(p = 0.6)
}
