% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_multiple_comparisons}
\alias{ss_multiple_comparisons}
\title{Multiple Comparisons Sample Size}
\usage{
ss_multiple_comparisons(k, delta, sd, power = 0.8, alpha = 0.05)
}
\arguments{
\item{k}{Number of comparisons}

\item{delta}{Effect size}

\item{sd}{Standard deviation}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Overall significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size with Bonferroni correction
}
\examples{
ss_multiple_comparisons(k = 5, delta = 0.5, sd = 1)
}
