% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_precision}
\alias{ss_precision}
\title{Sample Size for Precision (Confidence Interval Width)}
\usage{
ss_precision(sd, width, conf_level = 0.95)
}
\arguments{
\item{sd}{Standard deviation}

\item{width}{Desired confidence interval width}

\item{conf_level}{Confidence level (default 0.95)}
}
\value{
List with sample size results
}
\description{
Calculate sample size based on desired precision
}
\examples{
ss_precision(sd = 10, width = 5)
}
