% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_prediction_binary}
\alias{ss_prediction_binary}
\title{Clinical Prediction Model Sample Size (Binary)}
\usage{
ss_prediction_binary(
  prevalence,
  cstatistic,
  parameters,
  shrinkage = 0.9,
  r2_pseudo = NULL
)
}
\arguments{
\item{prevalence}{Event prevalence}

\item{cstatistic}{Expected C-statistic (AUC)}

\item{parameters}{Number of candidate parameters}

\item{shrinkage}{Desired shrinkage (default 0.9)}

\item{r2_pseudo}{Expected pseudo R-squared (optional)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for developing binary prediction model
Based on Riley et al. methodology
}
\examples{
ss_prediction_binary(prevalence = 0.3, cstatistic = 0.75, parameters = 10)
}
