% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_prediction_continuous}
\alias{ss_prediction_continuous}
\title{Clinical Prediction Model Sample Size (Continuous)}
\usage{
ss_prediction_continuous(r2, parameters, shrinkage = 0.9)
}
\arguments{
\item{r2}{Expected R-squared}

\item{parameters}{Number of candidate parameters}

\item{shrinkage}{Desired shrinkage (default 0.9)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for developing continuous prediction model
}
\examples{
ss_prediction_continuous(r2 = 0.3, parameters = 8)
}
