% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_prediction_survival}
\alias{ss_prediction_survival}
\title{Clinical Prediction Model Sample Size (Survival)}
\usage{
ss_prediction_survival(cstatistic, parameters, p_cens = 0.5, shrinkage = 0.9)
}
\arguments{
\item{cstatistic}{Expected C-index}

\item{parameters}{Number of candidate parameters}

\item{p_cens}{Censoring proportion}

\item{shrinkage}{Desired shrinkage (default 0.9)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for developing survival prediction model
}
\examples{
ss_prediction_survival(cstatistic = 0.7, parameters = 10, p_cens = 0.5)
}
