% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_complete.R
\name{ss_repeated_measures}
\alias{ss_repeated_measures}
\title{Repeated Measures ANOVA Sample Size}
\usage{
ss_repeated_measures(
  delta,
  sd_within,
  n_timepoints,
  corr = 0.5,
  power = 0.8,
  alpha = 0.05
)
}
\arguments{
\item{delta}{Expected difference between groups}

\item{sd_within}{Within-subject standard deviation}

\item{n_timepoints}{Number of time points}

\item{corr}{Within-subject correlation (default 0.5)}

\item{power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
List with sample size results
}
\description{
Calculate sample size for repeated measures design
}
\examples{
ss_repeated_measures(delta = 3, sd_within = 5, n_timepoints = 4)
}
