% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{standardize_vars}
\alias{standardize_vars}
\title{Standardize Variables}
\usage{
standardize_vars(data, vars = NULL)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to standardize (NULL = all numeric)}
}
\value{
Data frame with standardized variables
}
\description{
Standardize numeric variables to mean=0, sd=1
}
\examples{
data(omni_glucose)
std_data <- standardize_vars(omni_glucose, vars = c("age", "bmi"))
}
