% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{stats_anova}
\alias{stats_anova}
\title{One-Way ANOVA}
\usage{
stats_anova(response, group)
}
\arguments{
\item{response}{Numeric response variable}

\item{group}{Factor grouping variable}
}
\value{
List with ANOVA results
}
\description{
Perform one-way analysis of variance
}
\examples{
data(omni_glucose)
stats_anova(omni_glucose$hba1c_3m, omni_glucose$treatment)
}
