% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{stats_chisq}
\alias{stats_chisq}
\title{Chi-Square Test}
\usage{
stats_chisq(x, y = NULL)
}
\arguments{
\item{x}{Factor or table}

\item{y}{Factor (optional, for 2-variable case)}
}
\value{
List with test results
}
\description{
Perform chi-square test of independence
}
\examples{
data(omni_glucose)
stats_chisq(omni_glucose$treatment, omni_glucose$complications)
}
