% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_icc}
\alias{stats_icc}
\title{Intraclass Correlation Coefficient}
\usage{
stats_icc(data, method = "oneway")
}
\arguments{
\item{data}{Data frame with measurements}

\item{method}{"oneway" or "twoway"}
}
\value{
List with ICC results
}
\description{
Calculate ICC for reliability
}
\examples{
\dontrun{
stats_icc(measurements_df)
}
}
