% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_kappa}
\alias{stats_kappa}
\title{Kappa Statistic}
\usage{
stats_kappa(rater1, rater2)
}
\arguments{
\item{rater1}{Rater 1 ratings}

\item{rater2}{Rater 2 ratings}
}
\value{
List with kappa results
}
\description{
Calculate Cohen's kappa for inter-rater agreement
}
\examples{
\dontrun{
stats_kappa(rater1, rater2)
}
}
