% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_km_survival}
\alias{stats_km_survival}
\title{Kaplan-Meier Survival Analysis}
\usage{
stats_km_survival(time, status, group = NULL)
}
\arguments{
\item{time}{Survival time}

\item{status}{Event status (1 = event, 0 = censored)}

\item{group}{Optional grouping variable}
}
\value{
List with KM estimates
}
\description{
Calculate Kaplan-Meier survival estimates
}
\examples{
data(omni_cancer)
stats_km_survival(omni_cancer$time, omni_cancer$status, omni_cancer$treatment)
}
