% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_logistic_regression}
\alias{stats_logistic_regression}
\title{Logistic Regression}
\usage{
stats_logistic_regression(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}
}
\value{
List with logistic regression results
}
\description{
Perform logistic regression analysis
}
\examples{
data(omni_glucose)
omni_glucose$poor_control <- ifelse(omni_glucose$hba1c_3m > 7, 1, 0)
stats_logistic_regression(poor_control ~ age + bmi, data = omni_glucose)
}
