% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_logrank_test}
\alias{stats_logrank_test}
\title{Log-rank Test}
\usage{
stats_logrank_test(time, status, group)
}
\arguments{
\item{time}{Survival time}

\item{status}{Event status (1 = event, 0 = censored)}

\item{group}{Grouping variable}
}
\value{
List with log-rank test results
}
\description{
Compare survival curves between groups
}
\examples{
data(omni_cancer)
stats_logrank_test(omni_cancer$time, omni_cancer$status, omni_cancer$treatment)
}
