% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_mcnemar}
\alias{stats_mcnemar}
\title{McNemar's Test}
\usage{
stats_mcnemar(x, y)
}
\arguments{
\item{x}{Before treatment}

\item{y}{After treatment}
}
\value{
List with test results
}
\description{
Perform McNemar's test for paired proportions
}
\examples{
\dontrun{
stats_mcnemar(before, after)
}
}
