% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{stats_normality}
\alias{stats_normality}
\title{Normality Test}
\usage{
stats_normality(x, plot = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{plot}{Logical, show Q-Q plot (default FALSE)}
}
\value{
List with normality test results
}
\description{
Perform Shapiro-Wilk normality test with Q-Q plot option
}
\examples{
data(omni_glucose)
stats_normality(omni_glucose$hba1c_3m)
}
