% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_complete.R
\name{stats_poisson_regression}
\alias{stats_poisson_regression}
\title{Poisson Regression}
\usage{
stats_poisson_regression(formula, data, offset = NULL)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{offset}{Optional offset variable}
}
\value{
List with Poisson regression results
}
\description{
Perform Poisson regression for count data
}
\examples{
\dontrun{
stats_poisson_regression(counts ~ treatment + offset(log(time)), data = mydata)
}
}
