% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{stats_regression}
\alias{stats_regression}
\title{Linear Regression}
\usage{
stats_regression(formula, data)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}
}
\value{
List with regression results
}
\description{
Perform linear regression analysis
}
\examples{
data(omni_glucose)
stats_regression(hba1c_3m ~ age + bmi, data = omni_glucose)
}
