% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_functions.R
\name{stats_wilcoxon}
\alias{stats_wilcoxon}
\title{Wilcoxon Rank-Sum Test}
\usage{
stats_wilcoxon(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{Numeric vector for group 1}

\item{y}{Numeric vector for group 2}

\item{alternative}{"two.sided", "greater", or "less"}
}
\value{
List with test results
}
\description{
Perform Wilcoxon/Mann-Whitney rank-sum test for two independent samples
}
\examples{
data(omni_glucose)
g1 <- omni_glucose$hba1c_3m[omni_glucose$treatment == "胰岛素"]
g2 <- omni_glucose$hba1c_3m[omni_glucose$treatment == "口服药"]
stats_wilcoxon(g1, g2)
}
