% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_complete.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize Variables}
\usage{
winsorize(x, probs = c(0.05, 0.95))
}
\arguments{
\item{x}{Numeric vector}

\item{probs}{Quantile limits (default c(0.05, 0.95))}
}
\value{
Winsorized vector
}
\description{
Winsorize extreme values
}
\examples{
winsorize(rnorm(100))
}
